--- .github/workflows/build.yml
@@ -1,34 +1,0 @@
-name: Build RMusic
-on:
-  - push
-  - pull_request
-jobs:
-  build:
-    strategy:
-      matrix:
-        java:
-          - 17
-        os:
-          - ubuntu-latest
-    runs-on: ${{ matrix.os }}
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v2
-      - name: Validate gradle wrapper
-        uses: gradle/wrapper-validation-action@v1
-      - name: Setup jdk ${{ matrix.java }}
-        uses: actions/setup-java@v1
-        with:
-          java-version: ${{ matrix.java }}
-      - name: Make Gradle wrapper executable
-        if: ${{ runner.os != 'Windows' }}
-        run: chmod +x ./gradlew
-      - name: Build
-        run: ./gradlew build
-      - name: Upload build artifacts
-        if: ${{ runner.os == 'Linux' && matrix.java == '17' }}
-        uses: actions/upload-artifact@v2
-        with:
-          name: RMusic
-          path: build/libs/*.jar
-
