+++ .gitignore
@@ -116,6 +116,3 @@ run/
 
 # Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
 !gradle-wrapper.jar
-
-*.sqlite
-*.db
