+++ gradle.properties
@@ -4,11 +4,11 @@ org.gradle.jvmargs=-Xmx1G
 # check these on https://modmuss50.me/fabric.html
 minecraft_version=1.19.2
 yarn_mappings=1.19.2+build.28
-loader_version=0.14.10
+loader_version=0.14.9
 # Mod Properties
-mod_version=0.0.5
+mod_version=1.0-SNAPSHOT
 maven_group=cn.rtast
 archives_base_name=RMusic
 # Dependencies
 # check this on https://modmuss50.me/fabric.html
-fabric_version=0.72.0+1.19.2
+fabric_version=0.71.0+1.19.2
