+++ gradlew.bat
@@ -87,4 +87,3 @@ exit /b 1
 if "%OS%"=="Windows_NT" endlocal
 
 :omega
-
