+++ settings.gradle
@@ -7,4 +7,3 @@ pluginManagement {
         gradlePluginPortal()
     }
 }
-
