--- src/main/kotlin/cn/rtast/rmusic/data/config/ConfigModel.kt
@@ -1,6 +1,0 @@
-package cn.rtast.rmusic.data.config
-
-
-data class ConfigModel(
-    val api: List<Platform>
-)
