--- src/main/kotlin/cn/rtast/rmusic/data/config/Platform.kt
@@ -1,6 +1,0 @@
-package cn.rtast.rmusic.data.config
-
-data class Platform(
-    val name: String,
-    val api: String
-)
