--- src/main/kotlin/cn/rtast/rmusic/data/netease/ProfileModel.kt
@@ -1,10 +1,0 @@
-package cn.rtast.rmusic.data.netease
-
-
-data class ProfileModel(
-    val avatarUrl: String,
-    val id: Long,
-    val backgroundUrl: String,
-    val cookie: String,
-    val nickname: String
-)
