--- src/main/kotlin/cn/rtast/rmusic/data/netease/login/Account.kt
@@ -1,19 +1,0 @@
-package cn.rtast.rmusic.data.netease.login
-
-data class Account(
-    val anonimousUser: Boolean,
-    val ban: Int,
-    val baoyueVersion: Int,
-    val createTime: Long,
-    val donateVersion: Int,
-    val id: Long,
-    val salt: String,
-    val status: Int,
-    val tokenVersion: Int,
-    val type: Int,
-    val uninitialized: Boolean,
-    val userName: String,
-    val vipType: Int,
-    val viptypeVersion: Long,
-    val whitelistAuthority: Int
-)
