--- src/main/kotlin/cn/rtast/rmusic/data/netease/login/AvatarDetail.kt
@@ -1,7 +1,0 @@
-package cn.rtast.rmusic.data.netease.login
-
-data class AvatarDetail(
-    val identityIconUrl: String,
-    val identityLevel: Int,
-    val userType: Int
-)
