--- src/main/kotlin/cn/rtast/rmusic/data/netease/login/Binding.kt
@@ -1,13 +1,0 @@
-package cn.rtast.rmusic.data.netease.login
-
-data class Binding(
-    val bindingTime: Long,
-    val expired: Boolean,
-    val expiresIn: Int,
-    val id: Long,
-    val refreshTime: Int,
-    val tokenJsonStr: String,
-    val type: Int,
-    val url: String,
-    val userId: Long
-)
