--- src/main/kotlin/cn/rtast/rmusic/data/netease/login/Experts.kt
@@ -1,3 +1,0 @@
-package cn.rtast.rmusic.data.netease.login
-
-class Experts
