--- src/main/kotlin/cn/rtast/rmusic/data/netease/login/LoginResponseModel.kt
@@ -1,11 +1,0 @@
-package cn.rtast.rmusic.data.netease.login
-
-data class LoginResponseModel(
-    val account: Account,
-    val bindings: List<Binding>,
-    val code: Int,
-    val cookie: String,
-    val loginType: Int,
-    val profile: Profile,
-    val token: String
-)
