--- src/main/kotlin/cn/rtast/rmusic/data/netease/search/Album.kt
@@ -1,13 +1,0 @@
-package cn.rtast.rmusic.data.netease.search
-
-data class Album(
-    val artist: Artists,
-    val copyrightId: Int,
-    val id: Int,
-    val mark: Int,
-    val name: String,
-    val picId: Long,
-    val publishTime: Long,
-    val size: Int,
-    val status: Int
-)
