--- src/main/kotlin/cn/rtast/rmusic/data/netease/search/Artists.kt
@@ -1,14 +1,0 @@
-package cn.rtast.rmusic.data.netease.search
-
-data class Artists(
-    val albumSize: Int,
-    val alias: List<Any>,
-    val fansGroup: Any,
-    val id: Int,
-    val img1v1: Int,
-    val img1v1Url: String,
-    val name: String,
-    val picId: Int,
-    val picUrl: Any,
-    val trans: Any
-)
