--- src/main/kotlin/cn/rtast/rmusic/data/netease/search/Result.kt
@@ -1,7 +1,0 @@
-package cn.rtast.rmusic.data.netease.search
-
-data class Result(
-    val hasMore: Boolean,
-    val songCount: Int,
-    val songs: List<Song>
-)
