--- src/main/kotlin/cn/rtast/rmusic/data/netease/search/SearchResponseModel.kt
@@ -1,6 +1,0 @@
-package cn.rtast.rmusic.data.netease.search
-
-data class SearchResponseModel(
-    val code: Int,
-    val result: Result
-)
