--- src/main/kotlin/cn/rtast/rmusic/data/netease/url/FreeTimeTrialPrivilege.kt
@@ -1,8 +1,0 @@
-package cn.rtast.rmusic.data.netease.url
-
-data class FreeTimeTrialPrivilege(
-    val remainTime: Int,
-    val resConsumable: Boolean,
-    val type: Int,
-    val userConsumable: Boolean
-)
