--- src/main/kotlin/cn/rtast/rmusic/data/netease/url/FreeTrialPrivilege.kt
@@ -1,8 +1,0 @@
-package cn.rtast.rmusic.data.netease.url
-
-data class FreeTrialPrivilege(
-    val cannotListenReason: Any,
-    val listenType: Any,
-    val resConsumable: Boolean,
-    val userConsumable: Boolean
-)
