--- src/main/kotlin/cn/rtast/rmusic/data/netease/url/UrlModel.kt
@@ -1,6 +1,0 @@
-package cn.rtast.rmusic.data.netease.url
-
-data class UrlModel(
-    val code: Int,
-    val `data`: List<Data>
-)
