+++ src/main/kotlin/cn/rtast/rmusic/models/netease/Album.kt
@@ -1,0 +1,13 @@
+package cn.rtast.rmusic.models.netease
+
+data class Album(
+    val artist: Artist,
+    val copyrightId: Int,
+    val id: Int,
+    val mark: Int,
+    val name: String,
+    val picId: Long,
+    val publishTime: Long,
+    val size: Int,
+    val status: Int
+)
