+++ src/main/kotlin/cn/rtast/rmusic/models/netease/Artist.kt
@@ -1,0 +1,14 @@
+package cn.rtast.rmusic.models.netease
+
+data class Artist(
+    val albumSize: Int,
+    val alias: List<Any>,
+    val fansGroup: Any,
+    val id: Int,
+    val img1v1: Int,
+    val img1v1Url: String,
+    val name: String,
+    val picId: Int,
+    val picUrl: Any,
+    val trans: Any
+)
