+++ src/main/kotlin/cn/rtast/rmusic/models/netease/FreeTimeTrialPrivilege.kt
@@ -1,0 +1,8 @@
+package cn.rtast.rmusic.models.netease
+
+data class FreeTimeTrialPrivilege(
+    val remainTime: Int,
+    val resConsumable: Boolean,
+    val type: Int,
+    val userConsumable: Boolean
+)
