+++ src/main/kotlin/cn/rtast/rmusic/models/netease/FreeTrialPrivilege.kt
@@ -1,0 +1,8 @@
+package cn.rtast.rmusic.models.netease
+
+data class FreeTrialPrivilege(
+    val cannotListenReason: Any,
+    val listenType: Any,
+    val resConsumable: Boolean,
+    val userConsumable: Boolean
+)
