+++ src/main/kotlin/cn/rtast/rmusic/models/netease/Result.kt
@@ -1,0 +1,7 @@
+package cn.rtast.rmusic.models.netease
+
+data class Result(
+    val hasMore: Boolean,
+    val songCount: Int,
+    val songs: List<Song>
+)
