+++ src/main/kotlin/cn/rtast/rmusic/models/netease/SearchResultModel.kt
@@ -1,0 +1,6 @@
+package cn.rtast.rmusic.models.netease
+
+data class SearchResultModel(
+    val code: Int,
+    val result: Result
+)
