+++ src/main/kotlin/cn/rtast/rmusic/models/netease/Song.kt
@@ -1,0 +1,18 @@
+package cn.rtast.rmusic.models.netease
+
+data class Song(
+    val album: Album,
+    val alias: List<Any>,
+    val artists: List<Artist>,
+    val copyrightId: Int,
+    val duration: Int,
+    val fee: Int,
+    val ftype: Int,
+    val id: Int,
+    val mark: Int,
+    val mvid: Int,
+    val name: String,
+    val rUrl: Any,
+    val rtype: Int,
+    val status: Int
+)
