+++ src/main/kotlin/cn/rtast/rmusic/models/netease/SongUrlModel.kt
@@ -1,0 +1,6 @@
+package cn.rtast.rmusic.models.netease
+
+data class SongUrlModel(
+    val code: Int,
+    val data: List<Data>
+)
