--- src/main/resources/assets/rmusic/lang/zh_cn.json
@@ -1,21 +1,0 @@
-{
-  "rmusic.api.search.netease": "[RMusic]正在搜索 [ %s ] (网易云音乐).",
-  "rmusic.chat.play.click": "点击开始播放: [ %s ].",
-  "rmusic.chat.songinfo": "[RMusic]歌曲名: [ %s ] 作者: [ %s ] %s",
-  "rmusic.player.mute.off": "[RMusic]已恢复.",
-  "rmusic.player.mute.on": "[RMusic]已静音.",
-  "rmusic.player.notplaying": "[RMusic]当前没有播放音乐无法完成操作.",
-  "rmusic.player.pause": "[RMusic]暂停播放音乐.",
-  "rmusic.player.pause.off": "[RMusic]当前已暂停, 请使用/rmusic resume恢复播放.",
-  "rmusic.player.play": "[RMusic]正在播放歌曲: [ %s ].",
-  "rmusic.player.play.wait": "[RMusic]正在获取歌曲URL...",
-  "rmusic.player.resume": "[RMusic]继续播放.",
-  "rmusic.player.setvolume": "[RMusic]将音量大小设置为 %s (设置范围为0~2).",
-  "rmusic.player.stop": "[RMusic]已停止正在播放的音乐.",
-  "rmusic.session.netease.login.failure": "[RMusic]登录失败!请检查账号或密码是否正确!",
-  "rmusic.session.netease.login.success": "[RMusic]登陆成功!",
-  "rmusic.session.netease.login.wait": "[RMusic]正在登陆中...",
-  "rmusic.session.netease.logout.failure": "[RMusic]无法退出, 你还没有登录!",
-  "rmusic.session.netease.logout.success": "[RMusic]已退出!",
-  "rmusic.session.netease.logout.wait": "[RMusic]退出登陆中..."
-}
