+++ src/main/resources/rmusic.mixins.json
@@ -6,7 +6,6 @@
   "mixins": [
   ],
   "client": [
-    "SoundEventMixin"
   ],
   "injectors": {
     "defaultRequire": 1
